v2/web/requirements.txt
django==4.2.3
django[bcrypt]==4.2.3
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-hijack==3.4.1

arrow==1.2.3
crispy-bootstrap3==2022.1
gitpython==3.1.32
nested-lookup==0.2.25
psycopg2-binary==2.9.6
requests==2.31.0
